+++ docs/changelog.md
@@ -7,7 +7,8 @@
 - Fixed removing quests marked as destroy from questlog
 - Added new conditions: `QuestCompletedDate`, `QuestAcceptedToday`, `QuestCompletedToday`, `QuestNeverAccepted`, `QuestNeverCompleted`, 
 - Added new condition `SkillLevel` to check if the skill level met the minimum requirement 
-- Added new conditions `BuildingConstructed` to check if specified building is present on farm 
+- Added new condition `BuildingConstructed` to check if specified building is present on farm 
+
 - Added new conditions `CommunityCenterCompleted` to check whether community center completed or not
 - Added new conditions `FriendshipStatus` to check friendship status of specified character (Married, Engaged, Dating, Friendly)
 - Rename conditions: `Friendship` to `FriendshipLevel` to distinguish from `FriendshipStatus`