+++ docs/changelog.md
@@ -9,7 +9,8 @@
 - Added new condition `SkillLevel` to check if the skill level met the minimum requirement 
 - Added new condition `BuildingConstructed` to check if specified building is present on farm 
 
-- Added new conditions `CommunityCenterCompleted` to check whether community center completed or not
+- Added new condition `CommunityCenterCompleted` to check whether community center completed or not
+
 - Added new conditions `FriendshipStatus` to check friendship status of specified character (Married, Engaged, Dating, Friendly)
 - Rename conditions: `Friendship` to `FriendshipLevel` to distinguish from `FriendshipStatus`