+++ docs/changelog.md
@@ -8,7 +8,7 @@
 - Added new conditions: `QuestCompletedDate`, `QuestAcceptedToday`, `QuestCompletedToday`, `QuestNeverAccepted`, `QuestNeverCompleted`, 
 - Added new condition `SkillLevel` to check if the skill level met the minimum requirement 
 - Added new condition `BuildingConstructed` to check if specified building is present on farm 
-- Added new condition `CommunityCenterCompleted` to check whether community center completed or not
+- Added new condition `IsCommunityCenterCompleted` to check whether community center completed or not
 - Added new condition `FriendshipStatus` to check friendship status of specified character (Married, Engaged, Dating, Friendly)
 - Rename conditions: `Friendship` to `FriendshipLevel` to distinguish from `FriendshipStatus`